backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.12.5
typing-extensions==4.15.0

tensorflow==2.16.1
keras==3.0.5
numpy==1.26.4

pydicom==2.4.3
SimpleITK==2.3.1
Pillow==10.1.0
scikit-image==0.21.0
opencv-python==4.8.1.78

matplotlib==3.7.2
scipy==1.11.3

pyvista==0.43.0
#pydantic_settings==2.12.0
curlify==2.2.1